+++ ExcelTableConverter.JiraTableConverter/JiraConverter.cs
@@ -1,4 +1,5 @@
 ﻿using System.Text;
+using System.Windows.Forms;
 using ExcelTableConverter.ExcelContent.Model;
 using ExcelTableConverter.JiraTableConverter.CellFormaters;
 using ExcelTableConverter.TableConverter;
@@ -8,10 +9,16 @@ namespace ExcelTableConverter.JiraTableConverter
 {
   public class JiraConverter : BaseTableConverter
   {
-    private ICellFormatter _cellFormatter;
+    private readonly IExtendedJiraFeatureModel _extendedFeatures;
+    private readonly ICellFormatter _cellFormatter;
 
-    public JiraConverter()
+    public JiraConverter() : this(new ExtendedJiraFeatureModel())
     {
+    }
+
+    internal JiraConverter(IExtendedJiraFeatureModel extendedFeatures)
+    {
+      _extendedFeatures = extendedFeatures;
       AddConverter(ConverterName(), this);
       _cellFormatter = new CellFormatter();
     }
@@ -31,13 +38,15 @@ namespace ExcelTableConverter.JiraTableConverter
       ArgumentUtility.EnsureNotNull(excelTable, "excelTable");
 
       StringBuilder str = new StringBuilder();
+      var isFirstRow = true;
       foreach (var row in excelTable.Rows)
       {
         foreach (var cell in row.Columns)
         {
-          str.Append(string.Format("| {0} ", PrepareCellValue(cell)));
+          str.Append(string.Format("{0} {1} ", GetColumnSeparator(isFirstRow), PrepareCellValue(cell)));
         }
-        str.AppendLine("|");
+        str.AppendLine(GetColumnSeparator(isFirstRow));
+        isFirstRow = false;
       }
       return str.ToString();
     }
@@ -47,9 +56,19 @@ namespace ExcelTableConverter.JiraTableConverter
       return _cellFormatter.Format(cell);
     }
 
+    private string GetColumnSeparator(bool isFirstRow)
+    {
+      return _extendedFeatures.FirstRowIsHeader && isFirstRow ? "||" : "|";
+    }
+
     public override string GetFileExtension()
     {
       return "txt";
+    }
+
+    public override UserControl ExtendedFeaturesUserControl
+    {
+      get { return _extendedFeatures.BoundExtendedFeaturesUserControl; }
     }
   }
 }