+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - "High quality table" border configuration for LaTeX table converter. (#3)
 - Bold and italic is considered in Jira table converter. (#10)
+- Jira extended feature: "First row is header" (#6)
 
 ### Changed
 - Convert dialog opens now in the center of the Excel window. (#9)